Creates command groups, moves checks & helpers to utils/
